Well, it's wzd.m4, based on recently written wzd-config.
This is my first expirience with m4, so you'd better to look at it becouse it
may contain some bugs and/or errors. For the same reason I didn't add this
to distro also.

Never the less, it seems to work:

[dd]
checking for wzd-config... /usr/local/bin/wzd-config
[dd]
checking for wzdftpd version >= 0.3.0... yes
[dd]

and it sets proper cflags and libs.