* feature: user cache system moved to lib
This is the first step in a quite big modification of the memory management.
All users and groups related functions will use caches (they are called at
least once per second and user !), and there will be no differences between
backend_storage and server_storage.

This change was motivated by the fact that the previous system was not
designed to use uid, but index, and made too many calls to backend. It was
also very difficult to use anything else than the plaintext backend.

Functions GetUserByID and GetUserByName now check in cache, then make query
directly to backend (removing one useless function call).

The next (easy) step is to do the same for groups.