Step 2 of the big user/group management change: a group cache was integrated to the lib, and all functions also uses this cache.
There remains one problem: we need to find a way to ask backend for the list
of all users (groups) in one query, or at least the list of uid (gid).

The next step will be to modify the plaintext backend to store users and
groups, and then we can remove all tests related to backend_storage.
The question will be: what to do with the (big) portion of shared memory used
to store users and groups for plaintext ? It's not needed anymore for the
server, but external applications like siteconfig need a way to get user
informations ..

The backend_find_group function prototype was changed to work like
backend_find_user.

Many corrections were also made in wzd_site_group.c, to call directly
GetGroupByName (however we heavily rely on the group cache system to avoid
allocating and freeing memory).