* feature: GnuTLS support
The working-but-not-finished GnuTLS patch. ./configure now try to detect
GnuTLS first, and then OpenSSL if not found.
The current gnutls patch is fully working, but some fonctionnalities are
missing (cipher algorithms choice, buffer size etc).

For the moment, it is not question of suppressing openssl, but we will try
to support both implementations for a while.