The "oh, no, I broke it again !" commit
Humm, sorry, I could not resist, the CVS version was working too well to be a
CVS version ;)

I decided to recode completely to way the server manages clients (contexts): it
was previously a static array, and its size was fixed at compile time and could
never be change after that.
The context list is now really a linked list, so adding new connections is
really dynamic, and its size is no more fixed.

Of course, the maximum number of connections is still restricted by the
'max_threads' parameter, but it is now a soft limitation, and it can be change
using server variables.

Advantages are:
* less memory use at server start (because memory will be allocated when needed)
* there is no more limit on the maximum number of connections
* maximum number of connections can be changed at runtime

Of course, this is obviously a big change, affecting many files and with some
possible side-effects.

You have beed warned ! :)