* replace inconsistant int/unsigned int with uid_t or gid_t
WARNING ! This is not only a cosmetic change, the behaviour has changed.
For ex, 'nobody' is no more listed in standard users.

* splitted plaintext backend to have a file in charge of reading/writing
users file. It will be used to have an option for plaintext backend, to
have one file per user/group in a directory instead of one big file.

* remove direct prints to stderr in backends: this must *absolutely*
be avoided, in release mode the fd corresponding to stderr can be
attributed to a connection with a client !

* Fixed some warnings