The "I-know-it-will-break-all-archs-at-once" commit.
The previous directory architecture for source code wasn't good, and the
separation between libwzd-core and wzdftpd wasn't clear. So I decided to
split the src/ directory in libwzd-core/ and wzdftpd/

And now ? We have to fix all architectures. The current commit is ok on
linux, but it's highly probable that even BSD versions won't compile
now.

In the next few days, a new directory should be added to libwzd-core/
(perhaps named "archs" or "config"), and all platform-dependant
definitions will be moved to this directory, in separate files.