* feature: added initial implementation for config file helper * feature: added more functions to wzd_string_t * doc: updated inline documentation
The config file current implementation is working correctly (see
test file), but still misses some fonctionalities, like reading
comments correctly, etc. It is also not optimal since we do not
use a hash table yet to access data.
There is a problem with multiple values (we can store them
separated with commas). We need helper functions to access them,
and to modify the config_parse_data function to be able to read
multi-line data.

This new code will replace both the config file parser and the old
(and not working) code to save config.