* bugfix: fixed errors in sha and pam authentication
Added {cert} and {pam} code when changing password.

These changes are effective only for the plaintext backend, as other backends
does not currently use checkpass() and changepass() methods for now.

Note: changing a password using the pam method will really change the password
for pam, so the corresponding user on the system will also be changed.