* bugfix: permissions not sorted on keyword search (thanks to dindinx)
The bug was detected accidentally, searching why the resulting lib was marked
as having an executable stack by gcc. The problem was caused by a local
function in hash.c, in chtbl_extract(), and another error was detected in this
function (wrong function called, causing incorrect sort order).
This commit also fixes the stack-is-executable problem (doesn't change anything
on x86, since a writable stack is always executable ..).